refactor create_review_user_resource
